app/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
celery
psycopg2
gunicorn
django-celery
django-celery-beat
feedparser==5.2.1
scrapy-djangoitem==1.1.1
